Allow the tower to hurt the enemy

+++ src/helpz/Constants.java
@@ -40,7 +40,7 @@ public class Constants {
 			return "";
 		}
 
-		public static float GetStartDmg(int towerType) {
+		public static int GetStartDmg(int towerType) {
 			switch(towerType) {
 				case CANNON:
 					return 25;
